backend/requirements.txt
asgiref==3.10.0
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
decorator==5.2.1
Flask==3.1.2
future==1.0.0
geocoder==1.38.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
numpy==2.3.4
python-dateutil==2.9.0.post0
python-whois==0.9.6
pytz==2025.2
ratelim==0.1.6
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
soupsieve==2.8
threadpoolctl==3.6.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
